Makefile Organization
Moved all Windows-specific makefiles and supporting scripts to build/windows.  When a build on windows is desired, running build.bat in the root directory will copy these files and then invoke make.  When a build on linux is desired, running build.sh in the root directory will perform a similar action for the linux makefiles.