MySQL Fix
Fixed a latent bug in the MySQL Data Provider, specifically in the Connection::set method -- which sets options on a MYSQL structure before connection.

The only thing that method is currently used for is setting the MYSQL_OPT_RECONNECT option, however the method takes an enum mysql_option parameter, and can therefore -- in theory -- be used to set any option prior to connection.

Due to an oversight, however, the enum mysql_option parameter was ignored, and MYSQL_OPT_RECONNECT was unconditionally passed through to the mysql_options function.

Due to the limited manner in which this function is currently used this caused no problems, as the function behaved as expected, however going forward this could've caused confusion and was not intended.