World Save Compression
Modified the World module's save and load routines to zlib deflate and inflate column data, respectively.  This reduces the amount of data saved in the backing store by about 80x in my tests.

Since communicating with the backing store -- whether a database or directly to disk -- is in most cases the bottleneck, this accelerates that process.  It also keeps the size of the world on disk from getting out of hand, as it was liable to previously.