Server Shutdown Changes
Cleaned up some shutdown-related code inside the Server class.

Previously it was possible for multiple concurrent shutdown operations to proceed simultaneously, with each one launching a worker thread, attempting to shut the server down, and then all but the first bailing out after discovering the server was already shut down.

Now each successive request to shutdown will see that a shutdown is in process, and bail out immediately, without the overhead of creating a worker thread.

Cleaned up the way in which locking/waiting is handled with respect to the shutdown process.