ClientList Fix
Fixed a race condition in the ClientList class:

Previously, when creating start/end iterators, the ClientList class did not acquire the read lock, leaving that for the ClientListIterator instance it was creating.

However, in invoking the ClientListIterator constructor, the begin or end method (as applicable) of the ClientList's internal std::unordered_set must be invoked to get iterators for the ClientListIterator to wrap.  It was therefore possible for the std::unordered_set to be modified between when these iterators were created and when the ClientListIterator instance acquired the read lock.

Accordingly, the creation of these iterators is now guarded by the read lock.