Windows Network Stack Fix
Fixed a race condition within the Windows network stack:

The Windows network stack uses the MCPP::NetworkImpl::ReferenceManager class internally, to ensure that asynchronous callbacks dispatched due to events on connected sockets do not try and access the ConnectionHandler after it has been destroyed.

Accordingly, the ConnectionHandler destructor waits on the RefereneceManager.

However, there was a race condition -- the ReferenceManager::End method did not properly wake the ReferenceManager's internal condition variable if the count variable reached zero -- and therefore if the ConnectionHandler destructor attempted to wait on the ReferenceManager while a reference was still held by an asynchronous callback, the destructor would be stuck waiting indefinitely.